+++ src/main/java/lotto/Application.java
@@ -7,6 +7,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
+import lotto.model.Lotto;
 
 public class Application {
     public static void main(String[] args) {
