+++ src/main/java/lotto/model/BonusNumber.java
@@ -1,0 +1,4 @@
+package lotto.model;
+
+public class BonusNumber {
+}
