+++ src/main/java/lotto/model/Lottos.java
@@ -1,0 +1,4 @@
+package lotto.model;
+
+public class Lottos {
+}
