+++ src/main/java/lotto/model/WinningNumber.java
@@ -1,0 +1,4 @@
+package lotto.model;
+
+public class WinningNumber {
+}
