+++ src/main/java/lotto/model/WinningRank.java
@@ -1,0 +1,4 @@
+package lotto.model;
+
+public enum WinningRank {
+}
