+++ src/main/java/lotto/model/WinningResult.java
@@ -1,0 +1,4 @@
+package lotto.model;
+
+public class WinningResult {
+}
