+++ src/main/java/lotto/model/WinningResults.java
@@ -1,0 +1,4 @@
+package lotto.model;
+
+public class WinningResults {
+}
