+++ src/main/java/lotto/service/LottoMachine.java
@@ -1,0 +1,4 @@
+package lotto.service;
+
+public class LottoMachine {
+}
