+++ src/main/java/lotto/view/OutputView.java
@@ -1,0 +1,4 @@
+package lotto.view;
+
+public class OutputView {
+}
