+++ src/test/java/lotto/controller/LottoControllerTest.java
@@ -1,0 +1,4 @@
+package lotto.controller;
+
+public class LottoControllerTest {
+}
