+++ src/test/java/lotto/model/BonusNumberTest.java
@@ -1,0 +1,4 @@
+package lotto.model;
+
+public class BonusNumberTest {
+}
