+++ src/test/java/lotto/model/LottosTest.java
@@ -1,0 +1,4 @@
+package lotto.model;
+
+public class LottosTest {
+}
