+++ src/test/java/lotto/model/WinningNumberTest.java
@@ -1,0 +1,4 @@
+package lotto.model;
+
+public class WinningNumberTest {
+}
