+++ src/test/java/lotto/model/WinningRankTest.java
@@ -1,0 +1,4 @@
+package lotto.model;
+
+public class WinningRankTest {
+}
