+++ src/test/java/lotto/model/WinningResultTest.java
@@ -1,0 +1,4 @@
+package lotto.model;
+
+public class WinningResultTest {
+}
