+++ src/test/java/lotto/model/WinningResultsTest.java
@@ -1,0 +1,4 @@
+package lotto.model;
+
+public class WinningResultsTest {
+}
