+++ src/test/java/lotto/service/LottoMachineTest.java
@@ -1,0 +1,4 @@
+package lotto.service;
+
+public class LottoMachineTest {
+}
